new email template added

+++ src/services/emails/send.ts
@@ -19,37 +19,59 @@ export class SendEmailService {
     const handlerbarsTemplate = makeHandlerbarsTemplateService();
     const template = await handlerbarsTemplate.parse(templateData);
 
-    const attachmentContent = fs.readFileSync(filePath).toString("base64");
-    const fileName = path.basename(filePath);
+    if (filePath) {
+      const attachmentContent = fs.readFileSync(filePath).toString("base64");
+      const fileName = path.basename(filePath);
 
-    await this.mailjet.post("send", { version: "v3.1" }).request({
-      Messages: [
-        {
-          From: {
-            Email: from?.email || (process.env.EMAIL_ADDRESS as string),
-            Name: from?.name || (process.env.NAME_ADDRESS as string),
+      await this.mailjet.post("send", { version: "v3.1" }).request({
+        Messages: [
+          {
+            From: {
+              Email: from?.email || (process.env.EMAIL_ADDRESS as string),
+              Name: from?.name || (process.env.NAME_ADDRESS as string),
+            },
+            To: [
+              {
+                Email: to.email,
+                Name: to.name,
+              },
+            ],
+            Subject: subject,
+            TemplateLanguage: true,
+            HTMLPart: template,
+            Attachments: [
+              {
+                ContentType:
+                  "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
+                Filename: fileName,
+                Base64Content: attachmentContent,
+              },
+            ],
           },
-          To: [
-            {
-              Email: to.email,
-              Name: to.name,
-            },
-          ],
-          Subject: subject,
-          TemplateLanguage: true,
-          HTMLPart: template,
-          Attachments: [
-            {
-              ContentType:
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-              Filename: fileName,
-              Base64Content: attachmentContent,
-            },
-          ],
-        },
-      ],
-    });
+        ],
+      });
 
-    fs.unlinkSync(filePath);
+      fs.unlinkSync(filePath); // Apaga o arquivo gerado após o envio
+    } else {
+      await this.mailjet.post("send", { version: "v3.1" }).request({
+        Messages: [
+          {
+            From: {
+              Email: from?.email || (process.env.EMAIL_ADDRESS as string),
+              Name: from?.name || (process.env.NAME_ADDRESS as string),
+            },
+            To: [
+              {
+                Email: to.email,
+                Name: to.name,
+              },
+            ],
+            Subject: subject,
+            TemplateLanguage: true,
+            HTMLPart: template,
+          },
+        ],
+      });
+    }
   }
 }
